remove code not behind feature flag
