+++ pkg/api/dtos/index.go
@@ -40,7 +40,6 @@ type IndexViewData struct {
 	// Empty means no auto-login redirect should occur.
 	AutoLoginRedirectURL  string `json:"autoLoginRedirectURL,omitempty"`
 	AssetSriChecksEnabled bool   `json:"-"`
-	NewPreferencesPage    bool   `json:"-"`
 	OFREPRootUrlEnabled   bool   `json:"-"`
 }
