+++ pkg/services/frontend/index.html
@@ -214,10 +214,6 @@
         window.__grafanaReduceBootdataAPI = true;
       [[end]]
 
-      [[if .NewPreferencesPage]]
-        window.__grafanaNewPreferencesPage = true;
-      [[end]]
-
       [[if .OFREPRootUrlEnabled]]
         window.__grafanaOFREPRootUrlEnabled = true;
       [[end]]
