+++ public/app/core/components/AppChrome/MegaMenu/hooks.ts
@@ -4,15 +4,10 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useLocation } from 'react-router-dom-v5-compat';
 import { useLocalStorage } from 'react-use';
 
-import {
-  useGetUserPreferencesQuery,
-  usePatchUserPreferencesMutation,
-} from '@grafana/api-clients/internal/rtkq/legacy/preferences/user';
 import { useListPreferencesQuery, useUpdatePreferencesMutation } from '@grafana/api-clients/rtkq/preferences/v1';
 import { type NavModelItem } from '@grafana/data';
 import { t } from '@grafana/i18n';
 import { reportInteraction } from '@grafana/runtime';
-import { useFlagGrafanaNewPreferencesPage } from '@grafana/runtime/internal';
 import { useGrafana } from 'app/core/context/GrafanaContext';
 import { useAppNotification } from 'app/core/copy/appNotification';
 import { setBookmark } from 'app/core/reducers/navBarTree';
@@ -39,19 +34,12 @@ import {
 } from './utils';
 
 export const usePinnedItems = () => {
-  const newPrefsEnabled = useFlagGrafanaNewPreferencesPage();
   const k8sPreferences = useListPreferencesQuery(
-    contextSrv.user.isSignedIn && newPrefsEnabled
-      ? { fieldSelector: `metadata.name=user-${contextSrv.user.uid}` }
-      : skipToken
-  );
-  // TODO remove the legacy query once newPrefsEnabled is fully rolled out
-  const legacyPreferences = useGetUserPreferencesQuery(
-    contextSrv.user.isSignedIn && !newPrefsEnabled ? undefined : skipToken
+    contextSrv.user.isSignedIn ? { fieldSelector: `metadata.name=user-${contextSrv.user.uid}` } : skipToken
   );
 
-  const preferences = newPrefsEnabled ? k8sPreferences.data?.items[0]?.spec : legacyPreferences.data;
-  const isLoading = newPrefsEnabled ? k8sPreferences.isLoading : legacyPreferences.isLoading;
+  const preferences = k8sPreferences.data?.items[0]?.spec;
+  const isLoading = k8sPreferences.isLoading;
   const pinnedItems = useMemo(() => preferences?.navbar?.bookmarkUrls || [], [preferences]);
 
   return { pinnedItems, isLoading };
@@ -112,8 +100,6 @@ const usePinning = ({
 } => {
   const dispatch = useDispatch();
   const notifyApp = useAppNotification();
-  const newPrefsEnabled = useFlagGrafanaNewPreferencesPage();
-  const [patchPreferences] = usePatchUserPreferencesMutation();
   const [patchPreferencesK8s] = useUpdatePreferencesMutation();
   const { pinnedItems, isLoading } = usePinnedItems();
 
@@ -130,14 +116,12 @@ const usePinning = ({
         onSuccess?.();
         return true;
       };
-      return newPrefsEnabled
-        ? patchPreferencesK8s({
-            name: `user-${contextSrv.user.uid}`,
-            patch: { spec: { navbar: { bookmarkUrls } } },
-          }).then(onResult)
-        : patchPreferences({ patchPrefsCmd: { navbar: { bookmarkUrls } } }).then(onResult);
+      return patchPreferencesK8s({
+        name: `user-${contextSrv.user.uid}`,
+        patch: { spec: { navbar: { bookmarkUrls } } },
+      }).then(onResult);
     },
-    [newPrefsEnabled, patchPreferences, patchPreferencesK8s, notifyApp]
+    [patchPreferencesK8s, notifyApp]
   );
 
   // Local copy of the pinned urls so pin/unpin updates the menu immediately (the patch mutation
