--- public/app/core/components/SharedPreferences/SharedPreferencesOld.tsx
@@ -1,265 +1,0 @@
-import { PureComponent } from 'react';
-import * as React from 'react';
-
-import { FeatureState } from '@grafana/data';
-import { selectors } from '@grafana/e2e-selectors';
-import { t, Trans } from '@grafana/i18n';
-import { reportInteraction } from '@grafana/runtime';
-import {
-  Button,
-  Field,
-  FieldSet,
-  Label,
-  TimeZonePicker,
-  WeekStartPicker,
-  FeatureBadge,
-  Combobox,
-  type ComboboxOption,
-  type WeekStart,
-  isWeekStart,
-} from '@grafana/ui';
-import { DashboardPicker } from 'app/core/components/Select/DashboardPicker';
-import { PreferencesService } from 'app/core/services/PreferencesService';
-import { changeTheme } from 'app/core/services/theme';
-
-import { getSelectableThemes } from '../ThemeSelector/getSelectableThemes';
-
-import { homeDashboardChanged } from './analytics/main';
-import { getLanguageOptions, getStyles, getTranslatedThemeName, type Props, type State } from './utils';
-
-class SharedPreferences extends PureComponent<Props, State> {
-  service: PreferencesService;
-  themeOptions: ComboboxOption[];
-  languageOptions: ComboboxOption[];
-  // Server value stashed at load so a save can tell whether the home dashboard actually changed.
-  initialHomeDashboardUID = '';
-
-  constructor(props: Props) {
-    super(props);
-
-    this.service = new PreferencesService(props.resourceUri);
-    this.state = {
-      isLoading: false,
-      isSubmitting: false,
-      theme: '',
-      timezone: '',
-      weekStart: '',
-      language: '',
-      queryHistory: { homeTab: '' },
-      navbar: { bookmarkUrls: [] },
-    };
-
-    const themes = getSelectableThemes();
-
-    // Options are translated, so must be called after init but call them
-    // in constructor to avoid memo-break of array changing every render
-    this.themeOptions = themes.map((theme) => ({
-      value: theme.id,
-      label: getTranslatedThemeName(theme),
-      group: theme.isExtra ? t('shared-preferences.theme.experimental', 'Experimental') : undefined,
-    }));
-    this.languageOptions = getLanguageOptions();
-
-    // Add default option
-    this.themeOptions.unshift({ value: '', label: t('shared-preferences.theme.default-label', 'Default') });
-  }
-
-  async componentDidMount() {
-    this.setState({
-      isLoading: true,
-    });
-    const prefs = await this.service.load();
-    this.initialHomeDashboardUID = prefs.homeDashboardUID ?? '';
-
-    this.setState({
-      isLoading: false,
-      homeDashboardUID: prefs.homeDashboardUID,
-      theme: prefs.theme,
-      timezone: prefs.timezone,
-      weekStart: prefs.weekStart,
-      language: prefs.language,
-      queryHistory: prefs.queryHistory,
-      navbar: prefs.navbar,
-    });
-  }
-
-  onSubmitForm = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const confirmationResult = this.props.onConfirm ? await this.props.onConfirm() : true;
-
-    if (confirmationResult) {
-      const { homeDashboardUID, theme, timezone, weekStart, language, queryHistory, navbar } = this.state;
-      reportInteraction('grafana_preferences_save_button_clicked', {
-        preferenceType: this.props.preferenceType,
-        theme,
-        language,
-      });
-      this.setState({ isSubmitting: true });
-      await this.service
-        .update({
-          homeDashboardUID,
-          theme,
-          timezone,
-          weekStart,
-          language,
-          queryHistory,
-          navbar,
-        })
-        .finally(() => {
-          this.setState({ isSubmitting: false });
-        });
-      const nextHomeDashboardUID = homeDashboardUID ?? '';
-      if (nextHomeDashboardUID !== this.initialHomeDashboardUID) {
-        homeDashboardChanged({
-          preferenceType: this.props.preferenceType,
-          action: nextHomeDashboardUID ? 'set' : 'cleared',
-        });
-      }
-
-      window.location.reload();
-    }
-  };
-
-  onThemeChanged = (value: ComboboxOption<string>) => {
-    this.setState({ theme: value.value });
-    reportInteraction('grafana_preferences_theme_changed', {
-      toTheme: value.value,
-      preferenceType: this.props.preferenceType,
-    });
-
-    if (value.value) {
-      changeTheme(value.value, true);
-    }
-  };
-
-  onTimeZoneChanged = (timezone?: string) => {
-    if (typeof timezone !== 'string') {
-      return;
-    }
-    this.setState({ timezone: timezone });
-  };
-
-  onWeekStartChanged = (weekStart?: WeekStart) => {
-    this.setState({ weekStart: weekStart ?? '' });
-  };
-
-  onHomeDashboardChanged = (dashboardUID: string) => {
-    this.setState({ homeDashboardUID: dashboardUID });
-  };
-
-  onLanguageChanged = (language: string) => {
-    this.setState({ language });
-
-    reportInteraction('grafana_preferences_language_changed', {
-      toLanguage: language,
-      preferenceType: this.props.preferenceType,
-    });
-  };
-
-  render() {
-    const { theme, timezone, weekStart, homeDashboardUID, language, isLoading, isSubmitting } = this.state;
-    const { disabled } = this.props;
-    const styles = getStyles();
-    const currentThemeOption = this.themeOptions.find((x) => x.value === theme) ?? this.themeOptions[0];
-
-    return (
-      <form onSubmit={this.onSubmitForm} className={styles.form}>
-        <FieldSet label={<Trans i18nKey="shared-preferences.title">Preferences</Trans>} disabled={disabled}>
-          <Field
-            loading={isLoading}
-            disabled={isLoading}
-            label={t('shared-preferences.fields.theme-label', 'Interface theme')}
-          >
-            <Combobox
-              options={this.themeOptions}
-              value={currentThemeOption.value}
-              onChange={this.onThemeChanged}
-              id="shared-preferences-theme-select"
-            />
-          </Field>
-
-          <Field
-            loading={isLoading}
-            disabled={isLoading}
-            label={
-              <Label htmlFor="home-dashboard-select">
-                <span className={styles.labelText}>
-                  <Trans i18nKey="shared-preferences.fields.home-dashboard-label">Home Dashboard</Trans>
-                </span>
-              </Label>
-            }
-            data-testid="User preferences home dashboard drop down"
-          >
-            <DashboardPicker
-              value={homeDashboardUID}
-              onChange={(v) => this.onHomeDashboardChanged(v?.uid ?? '')}
-              defaultOptions={true}
-              isClearable={true}
-              placeholder={t('shared-preferences.fields.home-dashboard-placeholder', 'Default dashboard')}
-              inputId="home-dashboard-select"
-            />
-          </Field>
-
-          <Field
-            loading={isLoading}
-            disabled={isLoading}
-            label={t('shared-dashboard.fields.timezone-label', 'Timezone')}
-            data-testid={selectors.components.TimeZonePicker.containerV2}
-          >
-            <TimeZonePicker
-              includeInternal={true}
-              value={timezone}
-              onChange={this.onTimeZoneChanged}
-              inputId="shared-preferences-timezone-picker"
-            />
-          </Field>
-
-          <Field
-            loading={isLoading}
-            disabled={isLoading}
-            label={t('shared-preferences.fields.week-start-label', 'Week start')}
-            data-testid={selectors.components.WeekStartPicker.containerV2}
-          >
-            <WeekStartPicker
-              value={weekStart && isWeekStart(weekStart) ? weekStart : undefined}
-              onChange={this.onWeekStartChanged}
-              inputId="shared-preferences-week-start-picker"
-            />
-          </Field>
-
-          <Field
-            loading={isLoading}
-            disabled={isLoading}
-            label={
-              <Label htmlFor="language-preference-select">
-                <span className={styles.labelText}>
-                  <Trans i18nKey="shared-preferences.fields.language-preference-label">Language</Trans>
-                </span>
-                <FeatureBadge featureState={FeatureState.preview} />
-              </Label>
-            }
-            data-testid="User preferences language drop down"
-          >
-            <Combobox
-              value={this.languageOptions.find((lang) => lang.value === language)?.value || ''}
-              onChange={(lang: ComboboxOption | null) => this.onLanguageChanged(lang?.value ?? '')}
-              options={this.languageOptions}
-              placeholder={t('shared-preferences.fields.language-preference-placeholder', 'Choose language')}
-              id="language-preference-select"
-            />
-          </Field>
-        </FieldSet>
-        <Button
-          disabled={isSubmitting}
-          type="submit"
-          variant="primary"
-          data-testid={selectors.components.UserProfile.preferencesSaveButton}
-        >
-          <Trans i18nKey="shared-preferences.save">Save preferences</Trans>
-        </Button>
-      </form>
-    );
-  }
-}
-
-export default SharedPreferences;
