+++ public/app/core/services/theme.test.ts
@@ -1,6 +1,5 @@
 import { generatedAPI as preferencesAPI } from '@grafana/api-clients/rtkq/preferences/v1';
 import { config } from '@grafana/runtime';
-import { FlagKeys } from '@grafana/runtime/internal';
 import { setTestFlags } from '@grafana/test-utils/unstable';
 
 import { backendSrv } from './backend_srv';
@@ -44,35 +43,21 @@ describe('changeTheme', () => {
     expect(preferencesAPI.endpoints.updatePreferences.initiate).not.toHaveBeenCalled();
   });
 
-  describe('when the newPreferencesPage flag is off', () => {
-    it('persists via the legacy preferences API', async () => {
-      await changeTheme('light', false);
-      expect(backendSrv.patch).toHaveBeenCalledWith('/api/user/preferences', { theme: 'light' });
-      expect(preferencesAPI.endpoints.updatePreferences.initiate).not.toHaveBeenCalled();
+  it('persists to the user resource via the k8s preferences API', async () => {
+    await changeTheme('light', false);
+    expect(preferencesAPI.endpoints.updatePreferences.initiate).toHaveBeenCalledWith({
+      name: 'user-abc123',
+      patch: { spec: { theme: 'light' } },
     });
+    expect(backendSrv.patch).not.toHaveBeenCalled();
   });
 
-  describe('when the newPreferencesPage flag is on', () => {
-    beforeEach(() => {
-      setTestFlags({ [FlagKeys.GrafanaNewPreferencesPage]: true });
-    });
-
-    it('persists to the user resource via the k8s preferences API', async () => {
-      await changeTheme('light', false);
-      expect(preferencesAPI.endpoints.updatePreferences.initiate).toHaveBeenCalledWith({
-        name: 'user-abc123',
-        patch: { spec: { theme: 'light' } },
-      });
-      expect(backendSrv.patch).not.toHaveBeenCalled();
-    });
-
-    it('falls back to the "user" resource name when the user has no uid', async () => {
-      contextSrv.user.uid = '';
-      await changeTheme('light', false);
-      expect(preferencesAPI.endpoints.updatePreferences.initiate).toHaveBeenCalledWith({
-        name: 'user',
-        patch: { spec: { theme: 'light' } },
-      });
+  it('falls back to the "user" resource name when the user has no uid', async () => {
+    contextSrv.user.uid = '';
+    await changeTheme('light', false);
+    expect(preferencesAPI.endpoints.updatePreferences.initiate).toHaveBeenCalledWith({
+      name: 'user',
+      patch: { spec: { theme: 'light' } },
     });
   });
 });
