+++ public/app/features/teams/TeamSettings.tsx
@@ -1,14 +1,14 @@
 import { useForm } from 'react-hook-form';
 
 import { Trans, t } from '@grafana/i18n';
-import { useFlagGrafanaNewPreferencesPage } from '@grafana/runtime/internal';
 import { Button, Field, FieldSet, Input, Stack } from '@grafana/ui';
 import { TeamRolePicker } from 'app/core/components/RolePicker/TeamRolePicker';
 import { useRoleOptions } from 'app/core/components/RolePicker/hooks';
-import { SharedPreferences } from 'app/core/components/SharedPreferences/SharedPreferences';
 import { contextSrv } from 'app/core/services/context_srv';
 import { AccessControlAction } from 'app/types/accessControl';
 import { type Team } from 'app/types/teams';
+
+import { SharedPreferences } from '../../core/components/SharedPreferences/SharedPreferences';
 
 import { useUpdateTeam } from './hooks';
 
@@ -45,8 +45,7 @@ const TeamSettings = ({ team }: Props) => {
       },
     });
   };
-  const newPrefsEnabled = useFlagGrafanaNewPreferencesPage();
-  const teamResourceUri = newPrefsEnabled ? `team-${team.uid}` : `teams/${team.id}`;
+  const teamResourceUri = `team-${team.uid}`;
 
   return (
     <Stack direction={'column'} gap={3}>
