+++ public/views/index.html
@@ -317,10 +317,6 @@
         window.__grafanaAssetSriChecksEnabled = true;
       [[end]]
 
-      [[if .NewPreferencesPage]]
-        window.__grafanaNewPreferencesPage = true;
-      [[end]]
-
       [[if .OFREPRootUrlEnabled]]
         window.__grafanaOFREPRootUrlEnabled = true;
       [[end]]
